Generate build info from environment variable (#3368)
