+++ src/bin/tikv-importer.rs
@@ -13,9 +13,8 @@
 
 #![feature(slice_patterns)]
 
-#[macro_use]
-extern crate clap;
 extern crate chrono;
+extern crate clap;
 extern crate fs2;
 #[cfg(feature = "mem-profiling")]
 extern crate jemallocator;
@@ -51,27 +50,11 @@ use clap::{App, Arg, ArgMatches};
 
 use tikv::config::TiKvConfig;
 use tikv::import::ImportKVServer;
-use tikv::util::{self as tikv_util, panic_hook};
+use tikv::util::panic_hook;
 
 fn main() {
-    let long_version: String = {
-        let (hash, branch, time, rust_ver) = tikv_util::build_info();
-        format!(
-            "\nRelease Version:   {}\
-             \nGit Commit Hash:   {}\
-             \nGit Commit Branch: {}\
-             \nUTC Build Time:    {}\
-             \nRust Version:      {}",
-            crate_version!(),
-            hash,
-            branch,
-            time,
-            rust_ver
-        )
-    };
-
     let matches = App::new("TiKV Importer")
-        .long_version(long_version.as_ref())
+        .long_version(util::tikv_version_info().as_ref())
         .author("PingCAP Inc. <info@pingcap.com>")
         .about("An import server for TiKV")
         .arg(
@@ -118,7 +101,7 @@ fn main() {
     init_log(&config);
     initial_metric(&config.metric, None);
 
-    tikv_util::print_tikv_info();
+    util::print_tikv_info();
     panic_hook::set_exit_hook(false);
     check_environment_variables();
